# book-network-ui/src/app/services/models/borrowed-book-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BorrowedBookResponse {
  authorName?: string;
  id?: number;
  isbn?: string;
  rate?: number;
  returnApproved?: boolean;
  returned?: boolean;
  synopsis?: string;
  title?: string;
}