# book.network/src/main/java/com/aliboo/book/book/BorrowedBookResponse.java
package com.aliboo.book.book;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder

public class BorrowedBookResponse {
    private Integer id;
    private String title;
    private String authorName;
    private String isbn;
    private String synopsis;
    private double rate;
    private boolean returned;//nshofo wsh rj3 lbook wl no
    private boolean returnApproved;

}